+++ helpers/validateBody.js
@@ -1,8 +1,10 @@
+const HttpError = require("../helpers/HttpError");
+
 const validateBody = (schema) => {
   return (req, res, next) => {
     const { error } = schema.validate(req.body);
     if (error) {
-      return res.status(400).json({ message: "missing required name field" });
+      next(HttpError(400, error.message));
     }
     next();
   };
