+++ utils/index.js
@@ -1,0 +1,5 @@
+const ctrlWrapper = require("./controllerWrapper");
+
+module.exports = {
+  ctrlWrapper,
+};
